Bug Fixes
A few routines were executing in the wrong places, causing some unexpected actions. I moved execution of the round incrementer and player rotator from the onPause method of the ScoreEntry activity to the onClickListener of that activities action button. I also moved the routines to set the Bid/Score switches from the onClickListeners in the ScoreSheet activity to the action button listeners of the Bid and Score entry activities.