Game Face!
I added a new Dialog activity called Game Face. It allows users to set up a set of photos of themselves that are then used on the Score Sheet to depict a player's ranking in the game. The activity asks for a happy face photo, a sad face photo, and a game face photo. The happy face is displayed on the Score Sheet when the player is in first place, or tied for first place. The sad face photo is displayed when the player is in last place. The game face photo is displayed when a player is ranked between first and last place.